# Use faster kernel for LayerNorm forward