--- a/llmc/layernorm.cuh
+++ b/llmc/layernorm.cuh
@@ -448,6 +448,7 @@ __global__ void __launch_bounds__(512, 2) // todo - any warnings on Turing with 
 void layernorm_forward(floatX* out, floatX* mean, floatX* rstd,
                        floatX* inp, const floatX* weight, const floatX* bias,
                        int B, int T, int C, cudaStream_t stream) {
+    NVTX_RANGE_FN();
     const int block_size = 256;
     int block_y = block_size / WARP_SIZE;
     const int N = B * T;