I ran kernel 5 under `/dev/cuda/` (`./layernorm_forward 5`) on both RTX 3090 and H100 systems and it's faster on both of them.

Numbers:
kernel 3, optimal block size on:
- RTX 3090 → 32 (689.11 GB/s)
- H100 → 128-512 (1191.38 GB/s)

kernel 5, optimal block size on:
- RTX 3090 → 256 (773.13 GB/s)
- H100 → 256 (1244.5 GB/s)

It also seems to be more predictable in terms of what the optimal block size is, whereas kernel 3 perf varies a lot between these 2 systems.

Note: again, as per Ahmdal's law we're calling layernorm forward only once in our forward pass, the rest we are using the residual fused kernels. But still, it's a low hanging fruit unless I am missing out on some context as to why we didn't use kernel 5 here? Omission or by design? Is it because the way we're computing variance in kernel 5? (numerical stability?).