@gordicaleksa 2024-06-16
Also, given that their caching profiles are the same (we use `cs` directive in both for `mean`, `rstd` and `out`), I don't expect e2e performance to be worse with the new kernel.

I did profile e2e run as well but the difference is too minor and the inter-run variations are large in comparison. Having said that It looks like it's slightly faster.