@gordicaleksa 2024-06-16
Ran `profile_gpt2cu.py` on RTX 3090 and got the following:

kernel 3:
![prof_1](https://github.com/karpathy/llm.c/assets/29271842/99182c8e-b3ee-43a7-9914-d7d5eb062310)

kernel 6 (latest one):
![prof_2](https://github.com/karpathy/llm.c/assets/29271842/4e27fa9a-3c60-408b-b698-f7ff3cbd6287)

we went from 1.28 ms (0.29%) -> 1.11 ms (0.25%) per training step. It ain't much but it's honest work. :D